Components in Swag-Labs POM
Is an abstraction that represents a distinct, reusable part of the user interface (UI) in the application. It encapsulates both the data (e.g., form fields values) and the actions (e.g., filling out a form, clicking buttons) related to that part of the UI. Components are used to interact with and verify behavior of specific UI elements.